Button Text and About Banner
 - Have just started with the banner tonight on the home page and I
changed the colour scheme as it looked a little garish. Let me know if
you think black and grey seemed to follow on nicely from the front
page! Also not sure how to have full black across the whole banner.
Will look for a solution for that.

- Just played around with the font on the button and moved it further
down from the title. LOOKS SO GREAT THOUGH!