Background, add image to title?, made resizing relative
I have blurred the background a little, moved the title a little more
central and the button on centre point, and also I have sorted out the
resizing so that the title stays put regardless of how the window is
resized. I also added the image to the folder, was thinking of putting
that behind the title? couldn’t quite work out how to do it though. The
home page looks great! I’m not sure about how to get the black to go
across though.